+++ Atlas/app/equipes/Equipe1/index.html
@@ -123,7 +123,7 @@
   <!-- Import de Chart.js -->
   <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
   <!-- Script principal (logique JS) -->
-  <script src="./app.js" type="text/javascript"></script>
+  <script src="./app.js" type="module"></script>
 
 </body>
 </html>
